pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research_bot"
version = "0.1.0"
description = "Research assistant bots for Yonsei Intelligence Networking Lab"
requires-python = ">=3.11"
dependencies = [
  "anthropic>=0.40.0",
  "requests>=2.32.0",
  "pyyaml>=6.0.2",
  "python-dotenv>=1.0.1",
  "click>=8.1.7",
  "rich>=13.9.0",
  "tenacity>=9.0.0",
]

[project.optional-dependencies]
ingest = [
  "pdfminer.six>=20240706",
  "lancedb>=0.13.0",
  "sentence-transformers>=3.2.0",
  "rank-bm25>=0.2.2",
  "lxml>=5.3.0",
]
web = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.7.0",
]

[project.scripts]
research-bot = "research_bot.cli:main"
download-papers = "research_bot.scripts.download_papers:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
